Edit the Readme.md
Corrected the case for the id in the suggested table to match the syntax in the code.

Edited the name wu22.sqlite to studentdb.sqlite3 to match the syntax in the code.

Added instructions for how to create the table in tablePlus or a similar program.

Also added a randomly generated group of names to populate the table.